diagnosticos_uim/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-tasks==2.*
google-cloud-storage==2.*
google-api-python-client==2.*
google-auth==2.*
protobuf==4.*
flask==2.*
requests==2.*
pandas==1.*
python-dotenv==0.*
openpyxl==3.*
reportlab==3.*
jinja2==3.*
weasyprint==56.*